utils improvements

+++ src/main/java/org/black/kotlin/utils/ProjectUtils.java
@@ -1,46 +1,36 @@
 package org.black.kotlin.utils;
 
-import com.google.common.base.Predicate;
-import com.google.common.base.Predicates;
-import com.google.common.collect.Lists;
-import com.google.common.collect.Sets;
-import java.nio.file.*;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.List;
-import java.util.Set;
+import javax.swing.JEditorPane;
 import org.black.kotlin.project.KotlinProject;
-import org.black.kotlin.project.KotlinProjectConstants;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.jetbrains.kotlin.load.kotlin.PackageClassUtils;
 import org.jetbrains.kotlin.name.FqName;
-import org.jetbrains.kotlin.psi.JetFile;
-import org.openide.filesystems.FileObject;
-import org.netbeans.api.project.Project;
-import org.netbeans.api.java.classpath.ClassPath;
-import org.netbeans.api.project.SourceGroup;
-import org.netbeans.api.project.Sources;
 import org.openide.DialogDisplayer;
 import org.openide.NotifyDescriptor;
+import org.openide.cookies.EditorCookie;
+import org.openide.filesystems.FileObject;
+import org.openide.nodes.Node;
 import org.openide.util.Exceptions;
-
+import org.openide.windows.TopComponent;
 
 /**
  *
  * @author Александр
  */
 public class ProjectUtils {
-    
+
     private static final String LIB_FOLDER = "lib";
     private static final String LIB_EXTENSION = "jar";
     public static final String KT_HOME = System.getenv("KT_HOME") + "\\";
-    
+
     private org.netbeans.api.project.ProjectUtils projUtils;
-    
+
     public static String findMain(FileObject[] files) throws FileNotFoundException, IOException {
         for (FileObject file : files) {
             if (!file.isFolder()) {
@@ -56,56 +46,56 @@ public class ProjectUtils {
                 }
             }
         }
-        
+
         return null;
     }
-    
-    public static String getOutputDir(KotlinProject proj){
+
+    public static String getOutputDir(KotlinProject proj) {
         
-        File path = new File(proj.getProjectDirectory().getPath()+"/build");
-        
-        if (!path.exists()){
+        File path = new File(proj.getProjectDirectory().getPath() + "/build");
+
+        if (!path.exists()) {
             path.mkdirs();
         }
-                
-        
-        String dir = proj.getProjectDirectory().getPath()+"/build";
+
+        String dir = proj.getProjectDirectory().getPath() + "/build";
         String[] dirs = dir.split("/");
         StringBuilder outputDir = new StringBuilder("");
-        
-        for (String str : dirs){
+
+        for (String str : dirs) {
             outputDir.append(str);
             outputDir.append("\\");
         }
         outputDir.append(proj.getProjectDirectory().getName()).append(".jar");
-        
+
         return outputDir.toString();
     }
 
-    
-    public static void clean(KotlinProject proj){
-        
+    public static void clean(KotlinProject proj) {
+
         try {
-            proj.getProjectDirectory().getFileObject("build").delete();
+            if (proj.getProjectDirectory().getFileObject("build") != null) {
+                proj.getProjectDirectory().getFileObject("build").delete();
+            }
         } catch (IOException ex) {
             Exceptions.printStackTrace(ex);
         }
-        
+
     }
-    
+
     @NotNull
-    public static List<String> getClasspath(){
+    public static List<String> getClasspath() {
         List<String> classpath = new ArrayList(KotlinClasspath.getKotlinClasspath());
-        
+
 //        for (String s: classpath){
 //            DialogDisplayer.getDefault().notify(new NotifyDescriptor.
 //                Message(s));
 //        }
-        
-        
         return classpath;
     }
+
     
+
     @Nullable
     public static String getPackageByFile(FileObject file) {
 //        JetFile jetFile = new JetFile(file,true);// = KotlinPsiManager.INSTANCE.getParsedFile(file);
@@ -115,7 +105,7 @@ public class ProjectUtils {
 //        return jetFile.getPackageFqName().asString();
         return "stub";
     }
-    
+
     public static FqName createPackageClassName(FileObject file) {
         String filePackage = getPackageByFile(file);
         if (filePackage == null) {
@@ -123,14 +113,13 @@ public class ProjectUtils {
         }
         return PackageClassUtils.getPackageClassFqName(new FqName(filePackage));
     }
-    
-    
+
     public static String buildLibPath(String libName) {
         return KT_HOME + buildLibName(libName);
     }
-    
+
     private static String buildLibName(String libName) {
         return LIB_FOLDER + "/" + libName + "." + LIB_EXTENSION;
     }
-        
+
 }
